+++ src/main/java/chc/tfm/udt/repositorios/JugadorDAO.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JugadorDAO extends JpaRepository<JugadorEntity, Integer> {
+public interface JugadorDAO extends JpaRepository<JugadorEntity, Long> {
     /**
      * Metodo creado para buscar por nombre... se pueden crear metodos para buscar por cualquier atributo de la entidad.
      * @param nombre
